+++ CHANGELOG.md
@@ -6,6 +6,7 @@ PDF417 Changelog
 
 * Added validation of options in renderers.
 * Fixed a bug in ImageRenderer where padding was not in bgColor, but white.
+* Upgraded Intrevention/Image to v2
 
 0.1.0 (2014-12-24)
 ------------------